+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## revdeprun 2.1.0
 
 ### Bug fixes
 
@@ -12,7 +12,7 @@
 ### Improvements
 
 - Apply an xfun monkey patch to parallelize `xfun:::download_tarball()`.
-  Set the number of workers to 5 to speed up revdep tarball downloads
+  The number of workers is set to 5 to speed up revdep tarball downloads
   during `xfun::rev_check()` (#127).
 
 ## revdeprun 2.0.0